CP-307 fix: refactor blog/common to typescript
